refactor to have database class
