+++ data_output/test.csv
@@ -1,3 +1,3 @@
 Highlight;Title;Author;URL;Note;Location;Date
-Test description 1;Test_author;Test_author;;Test Note 1;1;2023.02.25 23:58
-Test description 2;Test_author;Test_author;;Test Note 2;2;2023.02.25 23:59
+Test description 1;test;test;;Test Note 1;1;2023.02.25 23:58
+Test description 2;test;test;;Test Note 2;2;2023.02.25 23:59
