+++ notification/notify.py
@@ -9,76 +9,7 @@ from email.message import EmailMessage
 
 import requests
 from datetime import datetime
-
-from notion_integration.notion_processing import read_notion_authorization_information, retrive_last_idx_number
-
-class NotionDatabase:
-    def __init__(self) -> None:
-        """This class is used to connect to the notion database and retrieve the data"""
-        
-        self.notion_key, self.notion_database = read_notion_authorization_information()
-        self.url = f"https://api.notion.com/v1/databases/{self.notion_database}/query"
-        self.headers = {
-            "Authorization": f"Bearer {self.notion_key}",
-            "accept": "application/json",
-            "Notion-Version": "2022-06-28",
-            "content-type": "application/json"
-        }
-
-    def query_database_notion_for_random_row(self):
-        """This function is used to query notion database the random row and return response in json format"""
-
-        last_idx = retrive_last_idx_number(self.notion_database, self.headers)
-        idx_to_filter = random.randint(1, last_idx)
-
-        payload = {
-                    "page_size": 1,
-                    "filter": {
-                    "property": "Idx",
-                    "number": {
-                        "equals": idx_to_filter}
-                    }
-                }
-        
-        response = requests.post(self.url, json=payload, headers=self.headers)
-
-        return response.json()
-
-    def retrieve_content_from_response(self, response):
-        """This function is used to retrieve the relevant content and format it from the response"""
-
-        item_retrieved = response['results'][0]['properties']
-
-        #format date to human readable
-        date_string = item_retrieved['Date']['date']['start']
-        date_format = '%Y-%m-%dT%H:%M:%S.%f%z'
-        datetime_obj = datetime.strptime(date_string, date_format)
-        date_formatted = datetime_obj.strftime('%Y-%m-%d %H:%M')
-
-        #format book to human readable
-        book_unformatted = item_retrieved['Book']['rich_text'][0]['text']['content']
-        book_formatted = book_unformatted.replace('_',' ')
-
-        #format location
-        location_unformatted = item_retrieved['Location']['number']
-        location_formatted = str(location_unformatted)
-
-        row_dict = {
-            'Highlight': item_retrieved['Highlight']['rich_text'][0]['text']['content'],
-            'Book': book_formatted,
-            'Author': item_retrieved['Author']['rich_text'][0]['text']['content'],
-            'Location': location_formatted,
-            'Date': date_formatted,
-            'Note': item_retrieved['Note']['title'][0]['text']['content'],
-        }
-
-        return row_dict
-
-    def get_row_data(self):
-        response = self.query_database_notion_for_random_row()
-        data = self.retrieve_content_from_response(response)
-        return data
-    
+from notion_integration.notion_book_database import NotionDatabase
 
 def read_database(database_path):
     # Open the CSV file for reading
@@ -168,6 +99,6 @@ if __name__ == "__main__":
         data = read_database(database_path)
     else:
         notion_database = NotionDatabase()
-        data = notion_database.get_row_data()
+        data = notion_database.get_random_row_data()
     send_email(data, sender, receiver)
     # send_notification_popup(data) #TODO: select option or separete script to pop up
